requirements.txt
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.3
tqdm==4.65.0
sentence-transformers==2.2.2
faiss-cpu==1.7.3
spacy==3.5.0
nltk==3.8.1
scikit-learn==1.3.0
huggingface_hub==0.14.1
